Add webpack 5 support
